+++ tests/extendrtests/src/rust/src/lib.rs
@@ -14,8 +14,6 @@ mod memory_leaks;
 
 mod altrep;
 
-mod altlist;
-
 // Return string `"Hello world!"` to R.
 #[extendr]
 fn hello_world() -> &'static str {
@@ -368,5 +366,4 @@ extendr_module! {
     use raw_identifiers;
     use memory_leaks;
     use altrep;
-    use altlist;
 }
